* Made the root system saveable. Should save some time in calculations. * Changed CGroup.dimOfRep from an int to a long. * The main group only resets if its Dynkin diagram has changed.

+++ src/tan/leveldecomposition/ui/LevelDecomposition.java
@@ -88,7 +88,7 @@ public class LevelDecomposition extends javax.swing.JPanel
         {
             Class[] types = new Class []
             {
-                java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.Integer.class, java.lang.Integer.class, java.lang.Integer.class, java.lang.Integer.class, java.lang.Integer.class
+                java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.Integer.class, java.lang.Long.class, java.lang.Integer.class, java.lang.Integer.class, java.lang.Integer.class
             };
             boolean[] canEdit = new boolean []
             {
